+++ tests/unit/iris/test_from_iris.py
@@ -16,27 +16,9 @@ from unittest.mock import patch
 from iris.coords import DimCoord
 from iris.cube import Cube
 
+from tests import MonitoredArray
+
 from ncdata.iris import from_iris
-
-
-class MonitoredArray:
-    """
-    An array wrapper for monitoring dask deferred accesses.
-
-    Wraps a real array, and can be read (indexed), enabling it to be wrapped with
-    dask.array_from_array.  It then records the read operations performed on it.
-    """
-
-    def __init__(self, data):
-        self.dtype = data.dtype
-        self.shape = data.shape
-        self.ndim = data.ndim
-        self._data = data
-        self._accesses = []
-
-    def __getitem__(self, keys):
-        self._accesses.append(keys)
-        return self._data[keys]
 
 
 def sample_cube(data_array=None):
